+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/EncryptedTransferWithMemoTransaction.java
@@ -6,6 +6,10 @@ import com.concordium.sdk.types.AccountAddress;
 import com.concordium.sdk.types.Nonce;
 import lombok.*;
 
+/**
+ * @deprecated Encrypted transfers are deprecated and partially removed since protocol version 7
+ */
+@Deprecated
 @Getter
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
@@ -38,7 +42,9 @@ public class EncryptedTransferWithMemoTransaction extends AccountTransaction {
      * @return Initialized {@link EncryptedTransferWithMemoTransaction}.
      * @throws TransactionCreationException On failure to create the Transaction from input params.
      *                                      Ex when any of the input param is NULL.
+     * @deprecated Encrypted transfers are deprecated and partially removed since protocol version 7
      */
+    @Deprecated
     @Builder
     public static EncryptedTransferWithMemoTransaction from(
             final EncryptedAmountTransferData data,